Add a "keyid" parse_field in public key packets
This keyid is computed to help the developer to interpret the signatures that
are parsed later on. The computation of the keyid is done on-the-fly by the
parser because v4 public key keyid are generated by hashing parts of the binary
representation of the packet. Doing it from the json representation would be
very cumbersome.