Add an "eol" option to output_options to specify line break style.
JSON exporter was using "\n" at several locations in the code to pretty print values. This new options allows the configuration of the end-of-line (eol) marker.
This allows formatting of output in DoS/MacOS/Linux styles and also to do no line-break and all object printed on a single line