Bugfix in PKCS#1 signature verification: some cases were not handled correctly.
Sometimes, we may try to check a signature corresponding to the wrong modulus, leading
to a signature value too big to fit in the group. It should be discarded as a wrong
signature.

In case we do not know the hash algorithm, return false instead of an exception. This
might actually be subject to discussion.