Bugfix: rewrite Pkcs1.raw_verify function which allowed for signature forgeries.
Due to a laxist handling of the digest_ingo ASN.1 DER structure, several
signature forgery techniques could be applied. The current fix checks that
the structure is idempotent when parsed then dumped.

Some work is needed in the DER at the preprocessor/core library level to
handle DER really correctly.